Lagre at søker bekrefter i søknadsobjekter
